git-svn-id: svn://svn.code.sf.net/p/iup/iup/trunk@3382 2f08bc20-984a-4df8-b195-9b5d5d477537

+++ iup/srclua5/iuplua.c
@@ -208,7 +208,7 @@ int iuplua_dofile(lua_State *L, const char *filename)
 
 int iuplua_dostring(lua_State *L, const char *s, const char *name)
 {
-  int status = luaL_loadbuffer(L, s, (int)strlen(s), name);
+  int status = luaL_loadbuffer(L, s, strlen(s), name);
   if (status == LUA_OK)
     status = docall(L, 0, 0);
   return report(L, status);
